Add std-lib and classpath
std-lib was previously added as an external jar and so we couldn't have
the .classpath file in the repository because the path for std-lib
would have been differen for each user of the repository. Now std-lib
has been added as an internal jar which also allows for .classpath to
be commited.